Word align all access to MMIO

+++ common/proto.c
@@ -3,10 +3,10 @@
 
 #include "../../mta1-mkdf-qemu-priv/include/hw/riscv/mta1_mkdf_mem.h"
 
-volatile uint8_t *can_rx = (volatile uint8_t *)MTA1_MKDF_MMIO_UART_RX_STATUS;
-volatile uint8_t *rx = (volatile uint8_t *)MTA1_MKDF_MMIO_UART_RX_DATA;
-volatile uint8_t *can_tx = (volatile uint8_t *)MTA1_MKDF_MMIO_UART_TX_STATUS;
-volatile uint8_t *tx = (volatile uint8_t *)MTA1_MKDF_MMIO_UART_TX_DATA;
+volatile uint32_t *can_rx = (volatile uint32_t *)MTA1_MKDF_MMIO_UART_RX_STATUS;
+volatile uint32_t *rx = (volatile uint32_t *)MTA1_MKDF_MMIO_UART_RX_DATA;
+volatile uint32_t *can_tx = (volatile uint32_t *)MTA1_MKDF_MMIO_UART_TX_STATUS;
+volatile uint32_t *tx = (volatile uint32_t *)MTA1_MKDF_MMIO_UART_TX_DATA;
 
 uint8_t genhdr(uint8_t id, uint8_t endpoint, uint8_t status, enum cmdlen len)
 {
